refactor chat 코드 정리
채팅 관련 코드 정리
