+++ src/main/java/com/hermez/farrot/chat/chatmessage/dto/response/LatestMessageResponse.java
@@ -1,7 +1,6 @@
 package com.hermez.farrot.chat.chatmessage.dto.response;
 
 import com.hermez.farrot.chat.chatmessage.entity.ChatMessageType;
-import com.querydsl.core.annotations.QueryProjection;
 import java.time.LocalDateTime;
 import lombok.Builder;
