+++ src/main/java/com/hermez/farrot/chat/chatroom/dto/SelectOption.java
@@ -1,4 +1,4 @@
-package com.hermez.farrot.chat.chatroom.controller;
+package com.hermez.farrot.chat.chatroom.dto;
 
 public record SelectOption(String code,String displayOption) {
